More refactoring and removal of unused .hbs files
