+++ app/controllers/order/confirm_booking.js
@@ -23,6 +23,11 @@ export default Ember.Controller.extend(cancelOrderMixin, asyncTasksMixin, {
     return order.get("isOnlineOrder") && !this.get("cart.canCheckout");
   },
 
+  emptyCart() {
+    let order = this.get("order");
+    return order.get("isOnlineOrder") && this.get("cart.isEmpty");
+  },
+
   actions: {
     browseMore() {
       this.transitionToRoute("browse");
@@ -30,6 +35,10 @@ export default Ember.Controller.extend(cancelOrderMixin, asyncTasksMixin, {
 
     confirmBooking() {
       let order = this.get("order");
+
+      if (this.emptyCart()) {
+        return this.i18nAlert("cart_content.empty_cart", _.noop);
+      }
 
       if (this.badCart()) {
         return this.i18nAlert("items_not_available", _.noop);
