--- app/templates/order/_cancel_order_button.hbs
@@ -1,7 +1,0 @@
-<div class="row order_button">
-  <div class="small-4 columns">
-  </div>
-  <div class="small-8 columns cancel_order">
-    <a href="#" id="cancel_order" class="drop_off" {{action 'cancelOrderPopUp' order.id}}><u>{{t "order.cancel_order"}}</u></a>
-  </div>
-</div>
