--- app/templates/order/_ggv_form.hbs
@@ -1,196 +1,0 @@
-{{#book-van}}
-  <div class="row">
-    <div class="name collection">{{t "order.vehicle.delivery_date_time"}}</div>
-    <div class="collection-data">{{t "order.limited_crossroads_opening_days_and_time"}}</div>
-  </div>
-  <div class="row time_inputs" >
-   <div class=" custom_radio_buttons small-12 columns"  for={{dateSlot1}}>
-      <div class="radio-button small-1 columns">
-        {{radio-button
-        checked=dateSlot1
-        name='selectedDate'
-        id=dateSlot1
-        value=dateSlot1
-        selection=selectedDate
-        required='true' }}
-      </div>
-      <label class="font-awesome-margin fa fa-calendar calender_icon radio-button-text"  for={{dateSlot1}} ></label>
-      <label class="radio-button-text align-text" for={{dateSlot1}}>
-        {{dateSlot1}}
-      </label>
-    </div>
-    <div class=" custom_radio_buttons small-12 columns">
-      <div class="radio-button small-1 columns" for={{dateSlot2}}>
-        {{radio-button
-        name='selectedDate'
-        id=dateSlot2
-        value=dateSlot2
-        selection=selectedDate
-        required='true'}}
-      </div>
-      <label class="font-awesome-margin fa fa-calendar calender_icon radio-button-text" for={{dateSlot2}} ></label>
-      <label class="radio-button-text align-text" for={{dateSlot2}}>
-        {{dateSlot2}}
-      </label>
-    </div>
-    <div class="small-12 columns time_selector">
-      {{select-list
-      content=timeSlots
-      on-change=(action (mut selectedTime))
-      selectedValue=selectedTime
-      prompt=(t 'time')
-      selectArrow=false
-      required='true'
-      pattern="[0-9]{1,2}:(00|30)\s?(AM|PM)"}}
-      <i class="fa fa-clock-o time_icon"></i>
-    </div>
-  </div>
-  <section class="book-van">
-    <div class="row">
-      <div class="small-12 columns">
-        <div class="row">
-          <div class="name collection">{{t "gogovan.book_van.delivery_district"}}</div>
-        </div>
-        <div class="row location-dropdown">
-          <div class="small-6 columns territory-select">
-            {{select-list
-            content=territories
-            on-change=(action (mut selectedTerritory))
-            selectedValue=selectedTerritory
-            value=selectedTerritory
-            required='true' }}
-          </div>
-          <div class="small-6 columns">
-            {{select-list
-            content=districtsByTerritory
-            on-change=(action (mut selectedDistrict))
-            selectedValue=selectedDistrict
-            value=selectedDistrict
-            required='true' }}
-          </div>
-        </div>
-        <div class="row">
-          <div class="name collection">{{t "gogovan.book_van.vehicle"}}</div>
-        </div>
-        <div class="gogovan-req ui">
-          <ul class='radio-buttons'>
-            {{#each gogovanOptions as |option|}}
-            <li class="radio">
-              {{radio-button name="selectionTest" selection=selectedGogovanOption value=option.id id=option.specialId}}
-              <label class="radio-wrap small-only-text-center" for={{option.specialId}}>
-                {{option.name}}
-              </label>
-            </li>
-            {{/each}}
-          </ul>
-        </div>
-        <div class="row">
-          <div class="small-8 columns">
-            <div class="name collection">{{t "gogovan.book_van.requirements"}}</div>
-            <div class='checkbox'>
-              {{input type="checkbox" name="speakEnglish" checked=speakEnglish id='1' }}
-              <label class="radio-button-text" for='1'>
-              <span>{{t "gogovan.book_van.speak_english"}}</span>
-              </label>
-            </div>
-            <div class='checkbox'>
-              {{input type="checkbox" name="borrowTrolley" checked=borrowTrolley id='2' }}
-              <label class="radio-button-text" for='2'>
-              <span>
-                {{#if isSelectedVan}}
-                {{t "gogovan.book_van.van_trolly_cost"}}
-                {{else}}
-                {{t "gogovan.book_van.truck_trolly_cost"}}
-                {{/if}}
-              </span>
-              </label>
-            </div>
-            <div class='checkbox'>
-              {{input type="checkbox" name="porterage" checked=porterage id='3' }}
-              <label class="radio-button-text" for='3'>
-              <span>{{t "gogovan.book_van.porterage"}}</span>
-              </label>
-            </div>
-            {{#if isSelectedVan}}
-            <div class='checkbox'>
-              {{input type="checkbox" name="longerGoods" checked=longerGoods id='4' }}
-              <label class="radio-button-text" for='4'>
-              <span>{{t "gogovan.book_van.longer_goods"}}</span>
-              </label>
-            </div>
-            {{#if longerGoods}}
-            <div class="radio_selection">
-              <div class="custom_radio_buttons">
-                {{radio-button name="longerGoods" selection=longGoodSelection value="half" id="half"}}
-                <label class="radio-button-text" for="half">
-                  {{t "gogovan.book_van.longer_goods_6ft"}}
-                </label>
-              </div>
-              <div class="custom_radio_buttons">
-                {{radio-button name="longerGoods" selection=longGoodSelection value="full" id="full"}}
-                <label class="radio-button-text" for="full">
-                  {{t "gogovan.book_van.longer_goods_6ft_tall"}}
-                </label>
-              </div>
-            </div>
-            {{/if}}
-            {{/if}}
-          </div>
-          <div class="small-4 columns">
-            <div class="ggv-price-box">
-              <div class="row">
-                <div class="small-8 columns price-text">
-                  {{t "gogovan.book_van.best_price_estimate"}}
-                  <span>${{gogovanPrice}}</span>
-                </div>
-                <div class="small-4 columns ggv_logo">
-                  <img src="assets/images/gogo-van.jpg" class='delivery-image'/>
-                </div>
-              </div>
-              <div class="ggv_link">
-                <a href="https://www.gogovan.com.hk/en/pricing" target="_blank">{{t "gogovan.book_van.learn_about_prices"}}</a>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="row">
-          <div class="name collection">{{t "gogovan.book_van.delivery_contact_person"}}</div>
-          <div class="collection-data">
-            {{t "gogovan.book_van.booking_with_name"}}
-            <br/>
-            {{t "gogovan.book_van.driver_call_to_confirm"}}
-          </div>
-        </div>
-        <div class="row time_inputs">
-          <div class="small-6 columns">
-            {{#form-control}}
-              {{input value=userName name="userName" id="name" maxlength="50" required=true placeholder="Name"}}
-            {{/form-control}}
-          </div>
-          <div class="small-6 columns mobile">
-            {{#form-control}}
-              <span class="prefix">+852</span>
-              {{numeric-input value=mobilePhone name="mobilePhone" id="mobile" maxlength="8" required=true pattern=".{8,}"}}
-            {{/form-control}}
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
-  <div class="row order_button">
-    <div class="small-4 columns">
-      {{#link-to "cart"}}
-        <i class="fa fa-angle-left" aria-hidden="true"></i>
-        {{t "order.return_to_cart"}}
-      {{/link-to}}
-    </div>
-    <div class="small-8 columns">
-      <button id="submit_pin" class="button expand book_van" {{action 'bookGGVSchedule'}}>
-        {{t "order.continue_to_confirm_details"}}
-      </button>
-    </div>
-  </div>
-  {{#if order.isDraft}}
-    {{partial "order/cancel_order_button"}}
-  {{/if}}
-{{/book-van}}
