--- app/templates/order/_plan_order_transport.hbs
@@ -1,35 +1,0 @@
-<div class="transport-detail">
-  <div class="name">{{t "order.transport_details.transport_method"}}</div>
-
-  <div class="row custom_radio_buttons">
-    <div class="radio-button small-1 columns">
-      {{radio-button
-      name="selectReason"
-      selection=selectedId
-      value="self"
-      id="self"}}
-    </div>
-    <label class="radio-button-text small-11 columns" for="self">
-      {{t "order.transport_details.collect_with_own_vehicle"}}
-    </label>
-  </div>
-  <div class="row custom_radio_buttons">
-    <div class="radio-button small-1 columns">
-      {{radio-button
-      name="selectReason"
-      selection=selectedId
-      value="ggv"
-      id="ggv"}}
-    </div>
-    <label class="radio-button-text small-11 columns" for="ggv">
-      {{t "order.transport_details.send_item_with_driver"}}
-    </label>
-  </div>
-
-  {{#if isSelfSelected}}
-    {{partial "order/self_vehicle_form"}}
-  {{else}}
-    {{partial "order/ggv_form"}}
-  {{/if}}
-
-</div>
