--- app/templates/order/_self_vehicle_form.hbs
@@ -1,72 +1,0 @@
-{{#if availableDates}}
-  <div class="row">
-    <div class="name collection">
-      {{t "order.vehicle.collection_date_time"}}
-    </div>
-    <div class="collection-data">
-      {{t "order.limited_crossroads_opening_days_and_time"}}
-    </div>
-  </div>
-  <div class="row time_inputs" >
-    <div class=" custom_radio_buttons small-12 columns"  for={{dateSlot1}}>
-      <div class="radio-button small-1 columns">
-        {{radio-button
-        checked=dateSlot1
-        name='selectedDate'
-        id=dateSlot1
-        value=dateSlot1
-        selection=selectedDate
-        required='true' }}
-      </div>
-      <label class="font-awesome-margin fa fa-calendar calender_icon radio-button-text"  for={{dateSlot1}} ></label>
-      <label class="radio-button-text align-text" for={{dateSlot1}}>
-        {{dateSlot1}}
-      </label>
-    </div>
-    <div class=" custom_radio_buttons small-12 columns">
-      <div class="radio-button small-1 columns" for={{dateSlot2}}>
-        {{radio-button
-        name='selectedDate'
-        id=dateSlot2
-        value=dateSlot2
-        selection=selectedDate
-        required='true'}}
-      </div>
-      <label class="font-awesome-margin fa fa-calendar calender_icon radio-button-text" for={{dateSlot2}} ></label>
-      <label class="radio-button-text align-text" for={{dateSlot2}}>
-        {{dateSlot2}}
-      </label>
-    </div>
-    <div class="small-12 columns time_selector">
-      {{select-list
-      content=timeSlots
-      on-change=(action (mut selectedTime))
-      selectedValue=selectedTime
-      prompt=(t 'time')
-      selectArrow=false
-      required='true'
-      pattern="[0-9]{1,2}:(00|30)\s?(AM|PM)"}}
-      <i class="fa fa-clock-o time_icon"></i>
-    </div>
-  </div>
-  <div class="row order_button">
-    <div class="small-4 columns">
-      {{#link-to "cart"}}
-      <i class="fa fa-angle-left" aria-hidden="true"></i>
-      {{t "order.return_to_cart"}}
-      {{/link-to}}
-    </div>
-    <div class="small-8 columns">
-      <button id="submit_pin" class="button expand drop_off" {{action 'bookSchedule'}}>
-        {{t "order.continue_to_confirm_details"}}
-      </button>
-    </div>
-  </div>
-  {{#if order.isDraft}}
-    {{partial "order/cancel_order_button"}}
-  {{/if}}
-{{else}}
-  <div class="loading_container">
-    {{partial 'loading'}}
-  </div>
-{{/if}}
